this code is working but occaisionally the pg pool or databasefails and im not sure why
