--- __tests__/articleById.test.js
@@ -1,63 +1,0 @@
-const request = require("supertest");
-const app = require("../app");
-const db = require("../db/connection");
-const seed = require("../db/seeds/seed");
-const data = require("../db/data/test-data");
-
-beforeEach(() => seed(data));
-afterAll(() => db.end());
-
-describe("GET /api/articles/:article_id", () => {
-  it("returns 200 and with an article object with the correct properties.", () => {
-    return request(app)
-      .get("/api/articles/1")
-      .expect(200)
-      .then(({ body }) => {
-        const { article } = body;
-        expect(article).toBeInstanceOf(Object);
-        expect(article).toHaveProperty("article_id")
-        expect(article).toHaveProperty("title");
-        expect(article).toHaveProperty("topic");
-        expect(article).toHaveProperty("author");
-        expect(article).toHaveProperty("body");
-        expect(article).toHaveProperty("created_at");
-        expect(article).toHaveProperty("votes");
-        expect(article).toHaveProperty("article_img_url");
-        expect(article.article_id).toEqual(1)
-      });
-  });
-  it("returns and object where the values are the correct data type", () => {
-    return request(app)
-      .get("/api/articles/1")
-      .expect(200)
-      .then(({ body }) => {
-        expect(body.article).toEqual(
-            expect.objectContaining({
-            article_id: expect.any(Number),
-            title: expect.any(String),
-            body: expect.any(String),
-            author: expect.any(String),
-            created_at: expect.any(String),
-            votes: expect.any(Number),
-          })
-        )
-      })
-     
-  });
-  it("returns 404 if the article doesnt exist", () => {
-    return request(app)
-      .get("/api/articles/99999")
-      .expect(404)
-      .then(({ body }) => {
-        expect(body.msg).toBe("404: Article Not Found");
-      });
-  });
-  it("returns 400 if the ID is not a number", () => {
-    return request(app)
-      .get("/api/articles/not-a-number")
-      .expect(400)
-      .then(({ body }) => {
-        expect(body.msg).toBe("400: Bad Request");
-      });
-  });
-});
