--- __tests__/articleComments.test.js
@@ -1,42 +1,0 @@
-const request = require("supertest");
-const app = require("../app");
-const db = require("../db/connection");
-const seed = require("../db/seeds/seed");
-const data = require("../db/data/test-data");
-
-beforeEach(() => seed(data));
-afterAll(() => db.end());
-
-describe("GET /api/articles/:article_id/comments", () => {
-  it("returns 200 and an array of objects with the correct properties", () => {
-    return request(app)
-      .get("/api/articles/1/comments")
-      .expect(200)
-      .then(({ body }) => {
-        const { comments } = body;
-        expect(comments).toBeInstanceOf(Array);
-        expect(comments.length).toBe(11)
-        comments.forEach((comment) => {
-          expect(comment).toBeInstanceOf(Object);
-          expect(comment).toHaveProperty("comment_id");
-          expect(comment).toHaveProperty("votes");
-          expect(comment).toHaveProperty("created_at");
-          expect(comment).toHaveProperty("author");
-          expect(comment).toHaveProperty("body");
-          expect(comment).toHaveProperty("article_id");
-        });
-      });
-  });
-  it("returns the comment in descing order (most recent first)", () => {
-    return request(app)
-      .get("/api/articles/1/comments")
-      .expect(200)
-      .then(({ body }) => {
-        const { comments } = body;
-        const dateSortedComments = [...comments].sort(
-          (a, b) => new Date(a.created_at) - new Date(b.created_at)
-        );
-        expect(comments).toEqual(dateSortedComments);
-      });
-  });
-});
