+++ controllers/articles.controller.js
@@ -1,4 +1,17 @@
-const { fetchArticle, fetchAllArticles, fetchArticleComments } = require("../models/articles.model")
+const { fetchArticle, fetchAllArticles, fetchArticleComments, insertComment } = require("../models/articles.model")
+
+
+exports.postArticleComment = (request, response, next) => {
+    const { article_id } = request.params;
+    const { username, body } = request.body;
+  
+    insertComment(article_id, username, body)
+      .then((comment) => {
+        response.status(201).send({ comment });
+      })
+      .catch(next);
+    }
+
 
 
 exports.getArticle = (request, response, next) => {
@@ -7,9 +20,7 @@ exports.getArticle = (request, response, next) => {
     .then((article) => {
         response.status(200).send({ article })
     })
-    .catch((error) => {
-        next(error);
-      });
+    .catch(next)
 }
 
 
@@ -18,9 +29,7 @@ exports.getAllArticles = (request, response, next) => {
     .then((articles) => {
         response.status(200).send({ articles })
     })
-    .catch((error) => {
-        next(error);
-      });
+    .catch(next)
 }
 
 exports.getArticleComments = (request, response, next) => {
@@ -29,7 +38,5 @@ exports.getArticleComments = (request, response, next) => {
     .then((comments) => {
         response.status(200).send({ comments })
     })
-    .catch((error) => {
-        next(error);
-      });
+    .catch(next)
 }
